frontend modularise init [WIP]
